Explictly include the 'builder' module
